+++ modules/os2forms_nemid/src/Element/MitidChildCpr.php
@@ -1,0 +1,64 @@
+<?php
+
+namespace Drupal\os2forms_nemid\Element;
+
+use Drupal\Core\Form\FormStateInterface;
+
+/**
+ * Provides a 'os2forms_mitid_child_cpr'.
+ *
+ * @FormElement("os2forms_mitid_child_cpr")
+ *
+ * @see \Drupal\Core\Render\Element\FormElement
+ * @see https://api.drupal.org/api/drupal/core%21lib%21Drupal%21Core%21Render%21Element%21FormElement.php/class/FormElement
+ * @see \Drupal\Core\Render\Element\RenderElement
+ * @see https://api.drupal.org/api/drupal/namespace/Drupal%21Core%21Render%21Element
+ * @see \Drupal\os2forms_nemid\Element\NemidCpr
+ */
+class MitidChildCpr extends NemidElementBase {
+
+  /**
+   * {@inheritdoc}
+   */
+  public function getInfo() {
+    $class = get_class($this);
+    return parent::getInfo() + [
+      '#process' => [
+        [$class, 'processMitidChildCpr'],
+        [$class, 'processAjaxForm'],
+      ],
+      '#element_validate' => [
+        [$class, 'validateMitidChildCpr'],
+      ],
+      '#pre_render' => [
+        [$class, 'preRenderMitidChildCpr'],
+      ],
+      '#theme' => 'input__os2forms_mitid_child_cpr',
+    ];
+  }
+
+  /**
+   * Processes a 'os2forms_mitid_child_cpr' element.
+   */
+  public static function processMitidChildCpr(&$element, FormStateInterface $form_state, &$complete_form) {
+    // Here you can add and manipulate your element's properties and callbacks.
+    return $element;
+  }
+
+  /**
+   * Webform element validation handler for #type 'os2forms_mitid_child_cpr'.
+   */
+  public static function validateMitidChildCpr(&$element, FormStateInterface $form_state, &$complete_form) {
+    // Here you can add custom validation logic.
+  }
+
+  /**
+   * {@inheritdoc}
+   */
+  public static function preRenderMitidChildCpr(array $element) {
+    $element = parent::prerenderNemidElementBase($element);
+    static::setAttributes($element, ['form-text', 'os2forms-mitid-child-cpr']);
+    return $element;
+  }
+
+}