+++ modules/os2forms_nemid/src/Element/NemidChildrenRadios.php
@@ -17,4 +17,20 @@ use Drupal\Core\Render\Element\Radios;
  */
 class NemidChildrenRadios extends Radios {
 
+  /**
+   * {@inheritdoc}
+   */
+  public function getInfo() {
+    return parent::getInfo() + [
+      '#ajax' => [
+        'callback' => [MitidChildrenSelectAjaxBehaviour::class, 'mitidChildrenSelectAjax'],
+        'event' => 'change',
+        'progress' => [
+          'type' => 'throbber',
+          'message' => $this->t('Please wait...'),
+        ],
+      ],
+    ];
+  }
+
 }