+++ modules/os2forms_nemid/src/Plugin/WebformElement/MitidChildCpr.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Drupal\os2forms_nemid\Plugin\WebformElement;
+
+/**
+ * Provides a 'os2forms_mitid_child_cpr' element.
+ *
+ * @WebformElement(
+ *   id = "os2forms_mitid_child_cpr",
+ *   label = @Translation("MitID Child CPR"),
+ *   description = @Translation("Provides a MitID Child CPR element."),
+ *   category = @Translation("NemID"),
+ * )
+ */
+class MitidChildCpr extends MitidChildElementBase {
+
+}