+++ modules/os2forms_nemid/src/Plugin/WebformElement/MitidChildFloor.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Drupal\os2forms_nemid\Plugin\WebformElement;
+
+/**
+ * Provides a 'os2forms_mitid_child_floor' element.
+ *
+ * @WebformElement(
+ *   id = "os2forms_mitid_child_floor",
+ *   label = @Translation("MitID Child Floor"),
+ *   description = @Translation("Provides a MitID Child Floor element."),
+ *   category = @Translation("NemID"),
+ * )
+ */
+class MitidChildFloor extends MitidChildElementBase {
+
+}