+++ modules/os2forms_nemid/src/Plugin/WebformElement/MitidChildHouseNr.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Drupal\os2forms_nemid\Plugin\WebformElement;
+
+/**
+ * Provides a 'os2forms_mitid_child_house_nr' element.
+ *
+ * @WebformElement(
+ *   id = "os2forms_mitid_child_house_nr",
+ *   label = @Translation("MitID Child House Nr"),
+ *   description = @Translation("Provides a MitID Child House Nr element."),
+ *   category = @Translation("NemID"),
+ * )
+ */
+class MitidChildHouseNr extends MitidChildElementBase {
+
+}