+++ modules/os2forms_nemid/src/Plugin/WebformElement/MitidChildKommunekode.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Drupal\os2forms_nemid\Plugin\WebformElement;
+
+/**
+ * Provides a 'os2forms_mitid_child_kommunekode' element.
+ *
+ * @WebformElement(
+ *   id = "os2forms_mitid_child_kommunekode",
+ *   label = @Translation("MitID Child kommunekode"),
+ *   description = @Translation("Provides a MitID Child kommunekode element."),
+ *   category = @Translation("NemID"),
+ * )
+ */
+class MitidChildKommunekode extends MitidChildElementBase {
+
+}