+++ modules/os2forms_nemid/src/Plugin/WebformElement/MitidChildName.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Drupal\os2forms_nemid\Plugin\WebformElement;
+
+/**
+ * Provides a 'os2forms_mitid_child_name' element.
+ *
+ * @WebformElement(
+ *   id = "os2forms_mitid_child_name",
+ *   label = @Translation("MitID Child Name"),
+ *   description = @Translation("Provides a MitID Child Name element."),
+ *   category = @Translation("NemID"),
+ * )
+ */
+class MitidChildName extends MitidChildElementBase {
+
+}