+++ modules/os2forms_nemid/src/Plugin/WebformElement/MitidChildPostalCode.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Drupal\os2forms_nemid\Plugin\WebformElement;
+
+/**
+ * Provides a 'os2forms_mitid_child_postal_code' element.
+ *
+ * @WebformElement(
+ *   id = "os2forms_mitid_child_postal_code",
+ *   label = @Translation("MitID Child Postal Code"),
+ *   description = @Translation("Provides a MitID Child Postal Code element."),
+ *   category = @Translation("NemID"),
+ * )
+ */
+class MitidChildPostalCode extends MitidChildElementBase {
+
+}