+++ modules/os2forms_nemid/src/Plugin/WebformElement/MitidChildStreet.php
@@ -1,0 +1,17 @@
+<?php
+
+namespace Drupal\os2forms_nemid\Plugin\WebformElement;
+
+/**
+ * Provides a 'os2forms_mitid_child_street' element.
+ *
+ * @WebformElement(
+ *   id = "os2forms_mitid_child_street",
+ *   label = @Translation("MitID Child street"),
+ *   description = @Translation("Provides a MitID Child street element."),
+ *   category = @Translation("NemID"),
+ * )
+ */
+class MitidChildStreet extends MitidChildElementBase {
+
+}